requirements.txt
Django
django-autoslug==1.9.8
django-cors-headers==3.8.0
django-filter==21.1
django-rest-auth==0.9.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
gunicorn==20.0.4
